requirements.txt
numpy~=1.26.4
pandas~=2.2.3
cleanlab~=2.7.1
scipy~=1.14.1
scikit-learn~=1.6.0
pillow~=11.0.0
requests~=2.32.3
torch~=2.5.1
sentence-transformers~=3.3.1
transformers~=4.47.0
termcolor~=2.5.0
tqdm~=4.67.1
typing_extensions~=4.12.2
setuptools~=75.6.0

pyspellchecker~=0.8.1

language_tool_python~=2.8.1
spacy~=3.8.3
textstat~=0.7.4
